feat: [main] 상품 유효성 검증 로직 추가

+++ src/service/products.service.js
@@ -1,13 +1,25 @@
-import { ProductsRepository } from "../repository/products.repositroy.js";
+import { ProductsRepository } from '../repository/products.repositroy.js';
 
 export class ProductsService {
   productsRepository = new ProductsRepository();
 
   createProduct = async (title, description, userId) => {
+    if (!title) {
+      return res.status(400).json({
+        message: '제목 입력이 필요합니다.',
+      });
+    }
+
+    if (!description) {
+      return res.status(400).json({
+        message: '설명 입력이 필요합니다.',
+      });
+    }
+
     const createdPost = await this.productsRepository.createProduct(
       title,
       description,
-      userId
+      userId,
     );
 
     return {
@@ -44,6 +56,12 @@ export class ProductsService {
   getProductById = async (productId) => {
     const product = await this.productsRepository.getProductById(productId);
 
+    if (!product) {
+      return res.status(404).json({
+        message: '찾으시는 상품이 존재하지 않습니다.',
+      });
+    }
+
     return {
       productId: product.productId,
       title: product.title,
@@ -60,14 +78,41 @@ export class ProductsService {
     const product = await this.productsRepository.getProductById(productId);
 
     if (!product) {
-      throw new Error("원하시는 상품이 존재하지 않습니다.");
+      return res.status(404).json({
+        message: '수정하려는 상품이 존재하지 않습니다.',
+      });
+    }
+
+    if (!title && !description && !status) {
+      return res.status(400).json({
+        message: '수정 정보는 최소 한 가지 이상이어야 합니다.',
+      });
+    }
+
+    const isValidStatus = status
+      ? status === 'FOR_SALE' || status === 'SOLD_OUT'
+      : true;
+
+    if (!isValidStatus) {
+      return res.status(400).json({
+        message: '지원하지 않는 상태입니다. (status: FOR_SALE | SOLD_OUT)',
+      });
+    }
+
+    const { userId } = req.users;
+    const isProductOwner = product.UserId === userId;
+
+    if (!isProductOwner) {
+      return res.status(403).json({
+        message: '상품 수정 권한이 없습니다.',
+      });
     }
 
     const updatedProduct = await this.productsRepository.updateProduct(
       productId,
       title,
       description,
-      status
+      status,
     );
 
     return {
@@ -83,9 +128,22 @@ export class ProductsService {
 
   deleteProduct = async (productId) => {
     const product = await this.productsRepository.getProductById(productId);
+
     if (!product) {
-      throw new Error("원하시는 상품이 존재하지 않습니다.");
+      return res.status(404).json({
+        message: '삭제하시려는 상품이 존재하지 않습니다.',
+      });
     }
+
+    const { userId } = req.users;
+    const isProductOwner = product.UserId === userId;
+
+    if (!isProductOwner) {
+      return res.status(403).json({
+        message: '상품 삭제 권한이 없습니다.',
+      });
+    }
+
     await this.productsRepository.deleteProduct(productId);
 
     return {
